stack: Avoid measuring children against preferred request mode
Due to how transitions are implemented, a stack may have to allocate
more space to its children than the stack itself is given. So it has to
measure a child to find out how much space the child actually needs.

It's problematic to measure a widget against its preferred size request
mode for two main reasons:
1. It may be extremely slow, since it will commonly have to do (multiple
   layers of) binary searching.
2. Many widgets have a not-quite-strictly-correct measure implementation
   for the opposite size request mode. Indeed, oftentimes the only way
   to implement correct measurement is to do a binary search.

So instead, measure the child for its overall minimum size in one
orientation, and then for the minimum size for that size in the other
orientation, according to its preferred size request mode.

If the size allocated to the stack was not enough for its child, the old
logic would result in the child widget gaining more *width* to match the
stack's own height. The new logic will instead give the child more width
up to its global minimum width, and then more *height* to match the
width (or the other way around for width-for-height widgets). Neither of
the two strategies has a reason to end up in a "better" layout than the
other one, but the new one has the nice property of avoiding measuring
the children against their preferred size request mode.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>