widget: Optimize allocation checks
With G_ENABLE_DEBUG, gtk_widget_allocate () verifies that the allocated
size is acceptable for the widget by comparing it against the widget's
measurements. Specifically, it would measure the widget's minimum width
for the allocated height (and complain if the allocated width is less
than the minimum), and the minimum height for the allocated width (and
again complain if the allocated height is less than that).

For non-constant-size widgets, it's enough to check either one of the
two to ensure that the allocated size is acceptable, if we additionally
know that the allocated width/height is not smaller than the widget's
overall minimum width/height.

While doing both checks is useful to an extent as an assertion that the
widget's measurements in both orientations are consistent, it can be
extremely costly to measure a widget against its preferred size request
mode. Moreover, the checks in gtk_widget_allocate () are intended to
verify that the parent of the widget has allocated an acceptable size to
it, under the assumption that the widget's own measurement
implementation makes sense, and not to sanity-check the widget's own
measurement implementation.

So avoid measuring widgets against their preferred size request mode;
retain the checks that the allocated size is not smaller than the
overall minimum width/height, and that it is acceptable according to the
measurement in the widget's preferred mode.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>