widget: Optimize adjusting allocation
Commit 3742fabae040d914e6ae58edf31170a54a980f21
"widget: Don't overadjust allocations" made it so that we measure the
widget against its preferred size request mode in order to ensure we
don't end up overly reducing its allocation. This is correct, but slow.

Fortunately in the typical case we would not be over-adjusting the
allocation when using the widget's overall natual measurement. So
optimize the logic by measuring the widget in its preferred orientation
for the natural size, and only measuring it in the other orientation if
the size for the natural size is larger than the original allocation.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>