Refine gdk_should_use_portal()
Rename the function to gdk_display_should_use_portal().

We want to always use portals when we are in a sandbox.
Otherwise, it depends on whether the portal is available
and is a new enough version. Therefore, pass the portal
interface name and the required minimum version to
this function.

Update all callers.