a11y atspi: Send added/removed window as child arg
`gtk_at_spi_emit_children_changed` takes the added/removed
child as the 5th parameter, and the sender as the 6th
parameter.

`gtk_at_spi_root_child_changed` was passing these in
the wrong order.

As a result, the "object:children-changed:add" AT-SPI
signal was sent incorrectly.

Example event seen in Accerciser's Event Monitor without
this fix in place when starting the "Entry" -> "Completion"
example in gtk4-demo:

    50.5 object:children-changed:add(1, 0, [application | gtk4-demo])
        source: [application | gtk4-demo]
        application: [application | gtk4-demo]

The application itself is shown as both, the event source
and the new child.

With this commit in place, the new window/frame is now
shown as the new child as expected:

    1.0 object:children-changed:add(1, 0, [frame | Completion])
        source: [application | gtk4-demo]
        application: [application | gtk4-demo]

This also matches the behavior with GTK 3's gtk3-demo:

    91.6 object:children-changed:add(1, 0, [frame | Entry Completion])
        source: [application | gtk3-demo]
        application: [application | gtk3-demo]