ci: Remove MESON_TEST_MAX_PROCESSES
This variable was introduced to avoid spurious and flaky
issues where the CI would fail to spawn threads and exit.

I am confident this should be fixed by [1] so let's remove
this and see how it goes again.

MESON_TEST_MAX_PROCESSES in .gitlab-ci/run-tests.sh was left
in place however now it defaults to number_cpus rather than 1.

[1] https://gitlab.gnome.org/Infrastructure/ansible/-/merge_requests/27