gtkatspiroot: Fix invalid free() call
`g_variant_get ("as")` gives a heap-allocated `GVariantIter` pointer,
not a `GStrv`, so the free function being used here was wrong. However,
the `event_types` variable wasn’t actually read at all, so it’s better
to just drop it completely.

This fixes heap corruption seen in valgrind when using GTK with
assistive technologies (such as a screen reader) turned on.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>