gtkatspiroot: Delay some allocations until absolutely necessary
On half the code paths in this function, the `sender`/`event_name` won’t
be passed anywhere which requires ownership, so there’s no need to make
a copy of them by default. Only copy them when needed.

This fixes a leak of the `sender` on the control flow paths where it’s
not inserted into the hash table.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>