memoryformat: Use max_tasks
We now grab rows in chunks, and we can't make use of more threads
than the total number of rows, divided by the chunk size. This
helps reduce the threading overhead for small sizes, and makes the
parallel run code competitive with the single-threaded code in those
cases, until the size reacheds ~ 2000 pixels, when the threading
starts to be faster.

For data, see
https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/7965