ngl: Conform to buffer offset alignment requirements
When using glMapBufferRange() for our globals we need to make sure the offset is a
multiple of GL_UNIFORM_BUFFER_OFFSET_ALIGNMENT.

Since we've only been doing that for globals and GskGlobalsInstace is 128 bytes,
this is usually the case. But it is valid to require 256 bytes
alignment, and in those cases we used to fail.

The case that discovered this was an old card using nouveau, but it may
also affect some Applie hardware.

Fixes: b154c9caf4c7b93e528f3aef38ff58212c452fef
Fixes: #7180
Related: #7053