dataformat: Change behavior wrt YUV
Instead of declaring the dataformat as YUV or RGB, only declare the
fourcc as YUV or RGB by having a get_rgb_fourcc() and get_yuv_fourcc()
function.

This makes the dataformat a neutral description of the pixel format
again and gets colorstate data out.

The main de-facto benefit is that there can now be 2 fourccs for the
same enum value, one for RGB and one for YUV, which will be necessary
for DRM_FORMAT_AYUV/DRM_FORMAT_ARGB8888 and similar.

Note that for now formats are still declared as YUV internally because
the conversion functions still need that info, but that will be changed
next.