gpu: Add GSK_GPU_AS_IMAGE_EXACT_SIZE
This allows using gsk_gpu_get_node_as_image() from repeat nodes which
need correct sizing for the repeat sampler.

Previously, we would always render an offscreen, but that is wasteful,
in particular when the repeated node is a texture node.
And this happens in particular with the checkerboard background of
Nautilus' thumbnails.

Fixes: #7182
Related: nautilus#3067