gpu: Add occlusion optimization for container nodes
If we encounter a fully opaque container node, assume there's a child
node that covers the full node and try to not draw anything below.

The original goal was to catch the case where opaque content (like an
opaque thumbnail) was drawn with a CSS background (like a checkerboard
pattern for thumbnails), but that case does not work as long as rounded
corners are used, because while both background and content get clipped
using an identical rounded clip, it's 2 different nodes.
So there is never a container covering both.

This optimization does however catch black background behind videos and
it's cheap enough to implement. so I decided to keep it.