ci: Dump a bit more information
Bugs and fixes in mesa have been affecting GTK quite a bit, with
the new renderers. So make the ci info script dump out the version
of mesa (and vulkan, for good measure).

Having to download and run the image to find this information again
is a bit cumbersome, so lets just dump it in the ci logs.

Other packages can be added if they are of interest.