flowbox: Fix measuring natural width
I don't understand what this branch was intended to optimize. Perhaps it
was meant to check for nat_items (i.e. max_children_per_line) being 1
instead; but then the same optimization could be implemented at the
get_largest_aligned_line_length level.

In practice, this branch was getting taken, and the width for only a
single child was getting returned as (both minimal and) natural width.

Fix this by always taking the general branch, which seems to produce
sensible measurements. Setting max_children_per_line to 1 is a weird way
to use a flow box anyway; use GtkBox for that.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>