boxlayout: Don't always reserve height for baseline alignment
...when doing height-for-width as well. This follows up on the original
change made in commit 6b59c138b2252d4cc1609c3cd2bd1bf508b9202a.

Note that baseline alignment, when it is enabled, is unlikely to work
with widgets that trade height for width. This is because, while the
width request in is assumed to monotonically decrease as the available
height is increased (though even this turns out to not always be true),
the same can not be guaranteed about baselines, which would prevent the
binary-search based layout algorithm from discovering the optimal way to
distribute the available space between child widgets.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>