boxlayout: Measure a single variable-size child directly
Add a fast path for the very common case of having a single inconstant-
size child, possibly along with several constant-size children. In this
case, we can avoid doing the binary search, and just query the child
itself for its minimum size for the available size. As a bonus, in this
case we're measuring the child in the same orientation that the box is
being measured, and not the opposite orientation like the binary search
does. Since the box propagates its child's preferred size request mode,
ancestor widgets will hopefully prefer to measure the box in the same
orientation that the child prefers, and we're now measuring the child
for the same orientation instead of inverting it.

This is in many ways like "passing through" a measure call to the single
inconstant-size child, while subtracting the space required for other
constant-size children, and doing the usual aggregation of the results
in the end.

For the common case of many many layers of nested boxes of alternating
orientations which all add constant-size "decorations" around a central
wrapping label, this cuts down on the number of Pango layouts required
to lay things out dramatically. A reftest for such a case is included.

See these Fractal [0] and Paper Plane [1] issue reports about the severe
performance issues caused by the many layers of nested boxes performing
binary searching.

[0]: https://gitlab.gnome.org/GNOME/fractal/-/issues/1183
[1]: https://github.com/paper-plane-developers/paper-plane/issues/379

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>