boxlayout: Avoid measuring a sole inconstant expand child
When all of the following are true:
1. the layout is not homogeneous,
2. there is a single inconstant-size child,
3. all the constant-size children (if any) report minimum = natural
   along the box's orientation,
4. none of the constant-size children should be expanded,
5. the inconstant-size child should be expanded,

we can go ahead and allocate all of the space remaining after
subtracting constant-size children's sizes and any spacing to the
inconstant-size child. Notably, we don't even have to measure it, which
can be a serious win if we would be to measure it against its preferred
size request mode (e.g. width-for-height for a height-for-width widget
such as a GtkLabel).

This is an "allocation passthrough" optimization, mirroring the "measure
passthrough" optimization implemented in the previous commit: the
allocation given to the box is passed through to the inconstant-size
child after subtracting the space for its constant-size siblings. Unlike
the measure passthrough, this is only applicable under specific, yet
common enough, circumstances listed above.

The specific case this is meant to address is a horizontal box with some
constant-size children and an hexpand wrappable label (seen e.g. in
GtkTooltipWindow). With this patch, allocating the box avoids measuring
the label at all, while previously it would be measured width-for-
height.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>