ci: added android ci run for a packaged gtk-demo
This is more advanced than the basic build check introduced in parent.
Instead this makes full use of gtk-android-builder to build a complete
application bundle for gtk-demo similar to what is already available for
flatpak.

The framework this commit adds for building application packages should
be flexible enough to add the other demos later on.