scrolledwindow: Report constant-size in more cases
In a lot of cases, we either don't actually use our child's
measurements, or do use them, but don't propagate our for_size when
measuring the child. In those cases, we don't need to propagate the
child's preferred size request mode, we can report constant-size
instead.

As it's the case elsehwere, reporting constant-size greatly simplifies
and optimizes layout for ancestor widgets. In case of the scrolled
window however it's even more drastic than usual, since large parts of
an interface tend to be wrapped in a scrolled window. In the cases where
we now report constant-size even when our child isn't, the scrolled
window now acts like a *height-for-width propagation barrier*; meaning
that any instances of height-for-width layout happening inside the
window's content view are limited in scope by the window. For example,
if there is a GtkBox that has to run a binary search to measure itself
outside of the scrolled window, and one inside, this means that both
binary searches can be run independently, as opposed to running a full
inner binary search for each iteration of the outer one.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>