stack: Fix measuring during size interpolation
When interpolate-size is set, we interpolate the stack's size request
from the last allocated size to the newly visible child's size request,
while the transition is running. This hopefully causes ancestor widgets
to smoothly change the allocation of the stack, and the user sees a
pretty animation.

To keep the layout from breaking, we need to report consistent width-
for-height and height-for-width measurements throughout the transition.
Our parent can measure us passing any valid for_size, and we need to
keep the responses consistent. This is complicated by the fact that
while we know the current progress of the transition and which size
we're animating *from*, we don't know which size we're animating *to*.

We base our measurements on those of our child, the one we're animating
towards, with lerp applied on top. When we're interpolating size in both
orientations (as opposed to staying homogeneous in one of them), we must
also adjust the for_size before we pass it on to the child, in order to
"undo" the lerp we would have done when being measured in the opposite
orientation. This is similar to what GtkRevealer does.

This is notoriously tricky to get right (which took me more than a
year), in particular:
* we have to handle for_size == -1 specially,
* we have to handle progress == 0.0 specially,
* we have to make sure the result of inverse_lerp is reasonable, and
  handle the case when it's not,
* we have to round down the result of inverse_lerp, and round up the
  result of lerp.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>