layoutmanager: With wfh == hfw, prefer hfw
We generally much prefer height-for-width over width-for-height, since
the former is typically better supported and faster, and the latter is
known to cause issues, most notably with wrappable labels. So when we
have two children with differing preferences and so have to make an
arbitrary choice between the two, prefer height-for-width.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>