text: Don't produce fontless text nodes
Pango items can in fact come out of itemization without a font,
and pango limps along, trying to render hexboxes in this case.
For gsk, it seems much easier to just not generate anything in
this corner case, rather than trying to make everything we do
with the font NULL-safe.