x11: Deprecate all public backend api
This sends a clear message that the X11 backend is going away
eventually.

All internal callers have been updated to ignore the deprecations.

The reasons why we want to drop the X11 backend in GTK5 are:
- There is much less need to keep X11 support on the client-side,
  compared to the server-side (which is covered, with xwayland).
- The backend code is only very sporadically maintained and
  nobody wants to add new features to it.
- It adds another column to our platform support matrix, and
  some of the cases are complicated, such as Vulkan / X11 or
  EGL-vs-GLX with its visual complications.
- Even if GTK5 loses its X11 backend, your GTK4 applications
  will still have it.

NOTE: There is *no* need to act on this deprecation immediately,
the X11 backend will continue to work as before for the entire
lifetime of GTK 4.