sizerequest: Measure overall size of single-orientation size group peers
The previous logic would measure all the peers in a size group for the
same size as the original widget being measured. This is broken,
because:

1. We would try to measure other peers for sizes that are less than what
   they actually need.

2. It causes contradictory measurements. Consider the following typical
   scenario: there is a size group that forces a bunch of widgets to
   report the same width (GTK_SIZE_GROUP_HORIZONTAL), and we're looking
   at one of its members. Its parent widget measures its minimal width,
   and gets the maximum of the minimum widths of all the size group
   members, let's call this width W. The parent then measures the height
   for the width of W, and since the group has no influence on heights,
   it is the actual minimum height for W of the specific widget that is
   returned, let's call it H. The parent then measures the widget
   horizontally for H, expecting to get W as the minimum width, since W
   is the overall minimum width, and W x H is an acceptable size. But
   instead it gets the maximum among the minimum widths of the members
   of the group when measured for H, which can be a lot larger than W.

3. for_size passed to one peer in a group has no relevance for other
   peers, since we cannot assume that their parent containers will just
   happen to allocate the same size in the opposite orientation to the
   widgets in the group.

Instead, measure overall size of peer widgets, while still propagating
for_size to the specific widget that is being measured. This results in
consistent measurements. Note that the resulting size requests won't
necesserily be equal (nor was this the case before this change), but it
seems to work well enough in practice.

For size groups in mode GTK_SIZE_GROUP_BOTH, we still propagate for_size
when measuring peers, since in that case it is a logical and consistent
thing to do.

Also, rename some variables for clarity.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>