sizerequest: Fix size group members falsely reporting constant-size
Unfortunately, when a widget that is interiorly constants-size is placed
in a GTK_SIZE_GROUP_BOTH size group together with an inconstant-size
peer, it can no longer be considered constant-size on the exterior,
since its measurements as returned from gtk_widget_measure () will
depend on the available size in the opposite orientation.

Detect this case, and return a common request mode among the widget's
peers.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>