macos: Native window controls
Support for native macOS window controls has been added
to the GtkWindowControls class.

A new class GtkWindowButtonsQuartz takes care of showing
window buttons on macOS.

I think it's okay to handle all window buttons from this
class, so spacing can be mac-like.
Also this gives a nice example on how to implement
similar decorations on Windows.

One more feature would be to disable the window controls if
the window is in "backdrop" state.

Anyway, this code shows native window controls can actually
work, and it's not *that* complicated.