gpu: Don't double clip edges
Due to using offscreens when drawing shadows, we are drawing twice:
Once into the offscreen and then the offscreen into the backbuffer.

We were using the node bounds rectangle for both of these operations
which caused the edge to be antialiased twice - if the edge was not
pixel-aligned.

This was a cause of the label-shadows.ui reftest randomly breaking on
some setups.

Test added.