text layout: Add an assertion
It is bad news if we end up in snapshot() with lines that are
(partially) invalid. For example, we may end up with the line
that contains the cursors having its cursors array purged,
causing us to not draw the blinking cursor despite us changing
cursor alpha in a tick callback.

Add an assertion to catch this when it happens.