css: Inherit font variations from settings
Compute the initial value for font-variation-settings from the
font variations that are part of the system default font, if any.

This has the same effect as if we had a @font-face rule in CSS
that combined the family and variations of the system default
font into a face that is then used by our CSS as the initial
font.