inspector: Align wrappable labels with xalign, not halign
Because of natural-wrap-mode being set to 'inherit' (the default) as
opposed to 'none', the label's natural width for a specific height
doesn't give the label enough space to display all the text without
wrapping. Since hexpand is also set, the label receives all the
available width, so this would not be an issue if it wasn't for halign,
which limits the allocation to the natural width.

Instead, use xalign, which shifts the Pango layout inside the label's
allocation. This gives the label a chance to make use of as much of the
available width as it wants.

Fixes https://gitlab.gnome.org/GNOME/gtk/-/issues/7242

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>