Deprecate the Broadway backend
This backend was an experiment when it was merged.

It never went beyond that, and it hasn't really been kept
up-to-date wrt. to changes in the rest of GDK and GSK.

And more importantly, nobody is actively maintaining it.

NOTE: There is *no* need to act on this deprecation immediately,
the Broadway backend will continue to exist for the entire
lifetime of GTK 4.