Avoid using deprecated GtkWindow::keys-changed internally
The GtkWindow::keys-changed signal was deprecated in GTK 4.10, but we're
still using it internally in GtkShortcutsWindow, and it'll generate
spurious deprecation warnings when using G_ENABLE_DIAGNOSTIC in the
environment.

Given that:

- the signal is used by GtkShortcutsWindow when associated to a
  GtkWindow, so that key changes will result in updated accelerator
  labels
- the association exists only when a GtkShortcutsWindow is used as a
  help overlay by GtkApplicationWindow

We can easily avoid the whole signal emission by using the signal's
default class closure of GtkApplicationWindow to notify of key changes
on GtkShortcutsWindow, which in turn will update the accelerator labels.