textview: Fix measuring height with anchored children
If we have priv->height already computed, that already takes any
anchored widgets into account, so we should not add their height again.
And when we don't have priv->height computed, we should still request at
least as much space as our anchored children need.

Of course, all of this is still wrong, in that:
* text view pretends to be constant-size, but it can be set up to wrap
  lines,
* anchored children are a part of the text layout and not just stacked
  on top of each other (and the text), so the bin-like MAX'ing of their
  sizes is only a conservative estimate of how much space the text view
  needs, and not an accurate answer in any way.

Fixes: https://gitlab.gnome.org/GNOME/gtk/-/issues/7257
Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>