application: Register host apps with portal
When not under a sandbox, register the app with the new host app
registry in XDG Desktop Portal. This allows host system apps be
better recognized, and behave a little better, when interacting
with portals.

This has to happen very early during initalization, before even
the gtk_init() call, because e.g. on Wayland the Settings portal
is used when opening the display.

See https://github.com/flatpak/xdg-desktop-portal/pull/1521