png: handle cICP chunk via libpng API
libpng 1.6.45 has gained the ability to read and write the cICP chunk.

libpng will now recognize the cICP chunk as a "known" chunk and thus will
not present it as a "user" chunk anymore, breaking our handling.
We thus need to switch to using the API.

The png data changes because libpng writes the cICP chunk at a different
position than we were, so adjust the tests reference data accordingly.

Fixes: #7271