gdk/contentserializer: Force specific URI schemes to serialize as URI
There is currently the issue that we want to convert the URI of files
provided via GVFS that have a representation in the filesystem (usually
via FUSE) into file:// URIs, that other applications that do not use
GIO/GVFS can read text/uri-list from the clipboard.

As a temporary solution, we have just whitelisted a set of URI schemas
whose files should directly be serialized as URI instead of the file://
conversion mechanism. I hope I can get GVFS to return TRUE from
GDaemonFile.is_native if the file has a mounted FUSE path (gvfs#258);
and to get GIO to drop the statement that FUSE files aren't native in
the doc comment of Gio.File.is_native.