im context: add invalid-composition signal
The new signal is emitted whenever a composition fails in any way, and
informs about the separate characters that were part of the composition
buffer up to that point.
The first handler's return value dictates whether the display should be
beeped.

This is only relevant for GtkIMContext implementations that are
derived from GtkIMContextSimple, since the other implementations
don't beep.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8956>