inspector: Don't show children & controllers twice
ed2716f9 incorrectly assumed that once a matching class function had
been found, the others wouldn't be checked and executed anymore.
That is not actually the case, so the child widgets and controllers of a
widget were added twice (once in the native_get_children and then again
in widget_get_children after it found the Gtk.Widget class function).

This is fixed by having the native_get_children only return the GDK
Surface and Renderer. Also the comment that is incorrectly pointing out
that we return after the first match was removed.

Fixes: #7310