scrolledwindow: Improve behavior if nested in other scrollables
Make the GtkScrolledWindow widget a better citizen in situations
where the widget might not be the only one in the event widget
stack that handles scroll events.

To do that, make scrolled windows remember whether scroll events
were being handled by this widget or not, and let events propagate
in the situations where it does not.

We naturally prefer the bubbling order when first choosing the
widget that does handle the events, although control during
scrolling is lent to the capture phase controller, in order to
ensure that child scrollables happening to move underneath the
pointer cannot take control away from scroll events.

If scrolling in the scrolledwindow moves past overshoot boundaries,
or if events happen towards a direction where the scrolled window
is not scrollable, the scrolled window widget will mark itself as
not handling scroll events any further, and letting those propagate
up to a next taker.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8926>